Поиск предложений в базе переделал с помощью OffersRepository.

+++ src/PrgData.Common/Orders/ClientOrderPosition.cs
@@ -57,7 +57,7 @@ namespace PrgData.Common.Orders
 
 		public decimal ServerCost { get; set; }
 
-		public ushort ServerQuantity { get; set; }
+		public uint ServerQuantity { get; set; }
 
 		public bool Duplicated { get; set; }
